dom/plugins/test/mochitest/test_queryContentsScaleFactorWindowed.html

<html>
<head>
  <title>NPAPI NPNVcontentsScaleFactor Test</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="utils.js"></script>
</head>

<body onload="runTests()">
  <script class="testbody" type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

    function runTests() {
      var pluginElement = document.getElementById("plugin");
      var contentsScaleFactor;
      var exceptionThrown = false;
      try {
        contentsScaleFactor = pluginElement.queryContentsScaleFactor();
      } catch (e) {
        exceptionThrown = true;
      }
      is(exceptionThrown, false, "Exception thrown getting contents scale factor.");
      is(isNaN(contentsScaleFactor), false, "Invalid return getting contents scale factor");
      ok(true, "Got Scale Factor of " + contentsScaleFactor);
      SimpleTest.finish();
    }
  </script>

  <embed id="plugin" type="application/x-test" width="400" height="400" wmode="window"></embed>
</body>
</html>
